Added checking product availability in stores

+++ src/main/java/com/bush/pharmacy_web_app/service/MedicineService.java
@@ -58,7 +58,7 @@ public class MedicineService {
                 .map(readMapper::map);
     }
 
-    public List<PharmacyBranchReadDto> findBranchesMedicineLocated(Integer medicineId) {
+    public List<PharmacyBranchReadDto> findBranchesMedicineLocated(Long medicineId) {
         return branchRepository.findBranchesMedicineLocated(medicineId)
                 .stream()
                 .map(branchReadMapper::map)
